Migrate purchase datetime to timestamp

+++ src/main/java/com/giraffes/tgbot/model/TransactionDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class TransactionDto {
-    private Integer datetime;
+    private Long datetime;
     private String transactionId;
     private String sender;
     private String value;
